Resize header
Since the upcoming payloads are going to be bigger than just one
kilobyte, the second field now takes the length of the payload as an
integer instead of a byte.

+++ src/Handler.java
@@ -15,7 +15,7 @@ public abstract class Handler {
     public int packetLength;
     public Header receivedHeader;
     public byte receivedPacketType;
-    public byte receivedPayloadLength;
+    public int receivedPayloadLength;
     public byte receivedStreamIdentifier;
     public byte[] receivedProducerIdentifier;
     public byte[] receivedPayloadLabel;
